+++ smoking-test/nested/Item.svelte
@@ -1,0 +1,3 @@
+<p>Hello svelte!</p>
+
+<style>p { color: green }</style>