+++ src/inspector.d.ts
@@ -1,0 +1,5 @@
+declare const defaultInspectorOptions: any
+declare const inspector: string
+declare const loader: string
+
+export { defaultInspectorOptions, inspector, loader }